Header file fix, Check csub.h game.h game2.h and the bottom of edwing.cpp to see example fo header working
