+++ csub.h
@@ -15,94 +15,23 @@
  *
  * =====================================================================================
  */
-//defined types
-typedef float Flt;
-typedef float Vec[3];
-typedef Flt	Matrix[4][4];
-//macros
-#define rnd() (((Flt)rand())/(Flt)RAND_MAX)
-#define random(a) (rand()%(a))
-#define VecZero(v) (v)[0]=0.0,(v)[1]=0.0,(v)[2]=0.0
-#define MakeVector(x, y, z, v) (v)[0]=(x),(v)[1]=(y),(v)[2]=(z)
-#define VecCopy(a,b) (b)[0]=(a)[0];(b)[1]=(a)[1];(b)[2]=(a)[2]
-#define VecDot(a,b)	((a)[0]*(b)[0]+(a)[1]*(b)[1]+(a)[2]*(b)[2])
-#define VecSub(a,b,c) (c)[0]=(a)[0]-(b)[0]; \
-							 (c)[1]=(a)[1]-(b)[1]; \
-(c)[2]=(a)[2]-(b)[2]
-//constants
-const float TIMESLICE = 1.0f;
-const float GRAVITY = -0.6f;
-#define PI 3.141592653589793
-#define ALPHA 1
-const int MAX_BULLETS = 1000;
-const Flt MINIMUM_ASTEROID_SIZE = 10.0;
-//-----------------------------------------------------------------------------
-//Joel's stuff
-#include <thread>
-extern void play_sound(ALuint a);
-
-//Setup timers
-const double OOBILLION = 1.0 / 1e9;
-extern struct timespec timeStart, timeCurrent;
-extern double timeDiff(struct timespec *start, struct timespec *end);
-extern void timeCopy(struct timespec *dest, struct timespec *source);
-//-----------------------------------------------------------------------------
-class Image {
-public:
-   int width, height;
-   unsigned char *data;
-   ~Image() { delete [] data; }
-   Image(const char *fname) {
-   if (fname[0] == '\0')
-   return;
-  //printf("fname **%s**\n", fname);
-   int ppmFlag = 0;
-   char name[40];
-  strcpy(name, fname);
-  int slen = strlen(name);
-   char ppmname[80];
-  if (strncmp(name+(slen-4), ".ppm", 4) == 0)
-   ppmFlag = 1;
-  if (ppmFlag) {
-   strcpy(ppmname, name);
-  } else {
-   name[slen-4] = '\0';
-   //printf("name **%s**\n", name);
-   sprintf(ppmname,"%s.ppm", name);
-  //printf("ppmname **%s**\n", ppmname);
-   char ts[100];
-   //system("convert eball.jpg eball.ppm");
-  sprintf(ts, "convert %s %s", fname, ppmname);
-   system(ts);
-   }
-  //sprintf(ts, "%s", name);
-   FILE *fpi = fopen(ppmname, "r");
-   if (fpi) {
-   char line[200];
-   fgets(line, 200, fpi);
-   fgets(line, 200, fpi);
-  //skip comments and blank lines
-   while (line[0] == '#' || strlen(line) < 2)
-   fgets(line, 200, fpi);
-  sscanf(line, "%i %i", &width, &height);
-   fgets(line, 200, fpi);
-  //get pixel data
-   int n = width * height * 3;
-   data = new unsigned char[n];
-   for (int i=0; i<n; i++)
-   data[i] = fgetc(fpi);
-   fclose(fpi);
-  } else {
-   printf("ERROR opening image: %s\n",ppmname);
-  exit(0);
-   }
-   if (!ppmFlag)
-   unlink(ppmname);
-  }
-};
-
-
-
+ #include <iostream>
+ #include <cstdlib>
+ #include <cstring>
+ #include <ctime>
+ #include <cmath>
+ #include <unistd.h>
+ #include <X11/Xlib.h>
+ //#include <X11/Xutil.h>
+ //#include <GL/gl.h>
+ //#include <GL/glu.h>
+ #include <X11/keysym.h>
+ #include <GL/glx.h>
+ //Audio library
+ #include <AL/alut.h>
+ #include <thread>
+ #include "log.h"
+ #include "fonts.h"
 class Global {
 	public:
 	    GLuint artTexture;
@@ -140,7 +69,7 @@ class Global {
 		GLuint rockTexture2;
 		GLuint bushTexture1;
 		GLuint bushTexture2;
-		
+
 	    //Store sound sources here.
 	    ALuint bulletSound;
 	    ALuint youDiedSound;
@@ -158,369 +87,4 @@ class Global {
 			yres = 900;
 			memset(keys, 0, 65536);
 		}
-} gl;
-class Ship {
-	public:
-		Vec dir;
-		Vec pos;
-		Vec vel;
-		float angle;
-		float color[3];
-	public:
-		Ship() {
-			VecZero(dir);
-			pos[0] = (Flt)(gl.xres/2);
-			pos[1] = (Flt)(gl.yres/2);
-			pos[2] = 0.0f;
-			VecZero(vel);
-			angle = 0.0;
-			color[0] = color[1] = color[2] = 1.0;
-		}
-};
-class Bullet {
-	public:
-		Vec pos;
-		Vec vel;
-		float color[3];
-		struct timespec time;
-	public:
-		Bullet() { }
-};
-class Asteroid {
-	public:
-		Vec pos;
-		Vec vel;
-		int health = 50;
-		int maxHp = 50;
-		int hpMissing = maxHp-health;
-		bool invuln = 0;
-		struct timespec invulnTimer;
-		int nverts;
-		int gunNum;
-		Flt radius;
-		Vec vert[8];
-		float angle;
-		float rotate;
-		float color[3];
-		struct Asteroid *prev;
-		struct Asteroid *next;
-	public:
-		Asteroid() {
-			prev = NULL;
-			next = NULL;
-		}
-		
-		void drawHealthBar(int x, int y){
-			glColor3f(0,1,0.5);
-			glPushMatrix();
-			glTranslatef(0, 0, 0);
-			glBegin(GL_QUAD_STRIP);
-			glVertex2f(x,y);
-			glVertex2f(x, y-10);
-			glVertex2f(x+50, y);
-			glVertex2f(x+50,y-10);
-			glEnd();
-			glPopMatrix();
-			//Current MISSING player HP
-			glColor3f(0.5f,0,0);
-			glPushMatrix();
-			glTranslatef(0, 0, 0);
-			glBegin(GL_QUAD_STRIP);
-			glVertex2f(x+50-hpMissing,y);
-			glVertex2f(x+50-hpMissing, y-10);
-			glVertex2f(x+50, y);
-			glVertex2f(x+50,y-10);
-			glEnd();
-			glPopMatrix();
-		}
-		
-		void checkInvuln(){
-			struct timespec invulnComparison;
-			clock_gettime(CLOCK_REALTIME, &invulnComparison);
-			double timeDifference = timeDiff(&invulnTimer, &invulnComparison);
-			//std::cout<<timeDifference<<std::endl;
-			if(timeDifference>0.5){
-				invuln=0;			
-			}
-		}
-};
-class Game {
-	public:
-		int astr_destroyed;
-		Ship ship;
-		Asteroid *ahead;
-		Bullet *barr;
-		Bullet *barrAst;
-		int nasteroids;
-		int nbullets;
-		int astBull;
-		struct timespec bulletTimer;
-		struct timespec enemyBulletTimer[3];
-		struct timespec itemTimer;
-		struct timespec mouseThrustTimer;
-		bool mouseThrustOn;
-	public:
-		Game() {
-			astr_destroyed=0;
-			ahead = NULL;
-			barr = new Bullet[MAX_BULLETS];
-			barrAst = new Bullet[MAX_BULLETS];
-			nasteroids = 0;
-			nbullets = 0;
-			astBull = 0;
-			mouseThrustOn = false;
-			//build 10 asteroids...
-			for (int j=0; j<2; j++) { //Change num asteroid
-				Asteroid *a = new Asteroid;
-				a->nverts = 8;
-				a->radius = rnd()*80.0 + 40.0;
-				//a->radius=10.0;
-				Flt r2 = a->radius / 2.0;
-				Flt angle = 0.0f;
-				Flt inc = (PI * 2.0) / (Flt)a->nverts;
-				for (int i=0; i<a->nverts; i++) {
-					a->vert[i][0] = sin(angle) * (r2 + rnd() * a->radius);
-					a->vert[i][1] = cos(angle) * (r2 + rnd() * a->radius);
-					angle += inc;
-				}
-				a->pos[0] = (Flt)(rand() % gl.xres);
-				a->pos[1] = (Flt)(rand() % gl.yres);
-				a->pos[2] = 0.0f;
-				a->angle = 0.0;
-				a->rotate = rnd() * 4.0 - 2.0;
-				a->color[0] = 0.8;
-				//a->color[1] = 0.8;
-				a->color[1] = 0.8;
-				a->color[2] = 0.7;
-				a->vel[0] = (Flt)(rnd()*2.0-1.0);
-				a->vel[1] = (Flt)(rnd()*2.0-1.0);
-				//std::cout << "asteroid" << std::endl;
-				//add to front of linked list
-				a->next = ahead;
-				if (ahead != NULL)
-					ahead->prev = a;
-				ahead = a;
-				++nasteroids;
-			}
-			clock_gettime(CLOCK_REALTIME, &bulletTimer);
-			clock_gettime(CLOCK_REALTIME, &enemyBulletTimer[3]);
-		}
-		~Game() {
-			delete [] barr;
-			delete [] barrAst;
-		}
-} g;
-//X Windows variables
-class X11_wrapper {
-	private:
-		Display *dpy;
-		Window win;
-		GLXContext glc;
-	public:
-		X11_wrapper() {
-			GLint att[] = { GLX_RGBA, GLX_DEPTH_SIZE, 24, GLX_DOUBLEBUFFER, None };
-			//GLint att[] = { GLX_RGBA, GLX_DEPTH_SIZE, 24, None };
-			XSetWindowAttributes swa;
-			setup_screen_res(gl.xres, gl.yres);
-			dpy = XOpenDisplay(NULL);
-			if (dpy == NULL) {
-				std::cout << "\n\tcannot connect to X server" << std::endl;
-				exit(EXIT_FAILURE);
-			}
-			Window root = DefaultRootWindow(dpy);
-			XVisualInfo *vi = glXChooseVisual(dpy, 0, att);
-			if (vi == NULL) {
-				std::cout << "\n\tno appropriate visual found\n" << std::endl;
-				exit(EXIT_FAILURE);
-			}
-			Colormap cmap = XCreateColormap(dpy, root, vi->visual, AllocNone);
-			swa.colormap = cmap;
-			swa.event_mask = ExposureMask | KeyPressMask | KeyReleaseMask |
-				PointerMotionMask | MotionNotify | ButtonPress | ButtonRelease |
-				StructureNotifyMask | SubstructureNotifyMask;
-			win = XCreateWindow(dpy, root, 0, 0, gl.xres, gl.yres, 0,
-					vi->depth, InputOutput, vi->visual,
-					CWColormap | CWEventMask, &swa);
-			set_title();
-			glc = glXCreateContext(dpy, vi, NULL, GL_TRUE);
-			glXMakeCurrent(dpy, win, glc);
-			show_mouse_cursor(0);
-		}
-		~X11_wrapper() {
-			XDestroyWindow(dpy, win);
-			XCloseDisplay(dpy);
-		}
-		void set_title() {
-			//Set the window title bar.
-			XMapWindow(dpy, win);
-			XStoreName(dpy, win, "CSUB BATTLE ROYALE!");
-		}
-		void check_resize(XEvent *e) {
-			//The ConfigureNotify is sent by the
-			//server if the window is resized.
-			if (e->type != ConfigureNotify)
-				return;
-			XConfigureEvent xce = e->xconfigure;
-			if (xce.width != gl.xres || xce.height != gl.yres) {
-				//Window size did change.
-				reshape_window(xce.width, xce.height);
-			}
-		}
-		void reshape_window(int width, int height) {
-			//window has been resized.
-			setup_screen_res(width, height);
-			glViewport(0, 0, (GLint)width, (GLint)height);
-			glMatrixMode(GL_PROJECTION); glLoadIdentity();
-			glMatrixMode(GL_MODELVIEW); glLoadIdentity();
-			glOrtho(0, gl.xres, 0, gl.yres, -1, 1);
-			set_title();
-		}
-		void setup_screen_res(const int w, const int h) {
-			gl.xres = w + 1000;
-			gl.yres = h + 1000;
-		}
-		void swapBuffers() {
-			glXSwapBuffers(dpy, win);
-		}
-		bool getXPending() {
-			return XPending(dpy);
-		}
-		XEvent getXNextEvent() {
-			XEvent e;
-			XNextEvent(dpy, &e);
-			return e;
-		}
-		void set_mouse_position(int x, int y) {
-			XWarpPointer(dpy, None, win, 0, 0, 0, 0, x, y);
-		}
-		void clear_screen() {
-			XClearWindow(dpy, win);
-		}
-		void show_mouse_cursor(const int onoff) {
-			if (onoff) {
-				//this removes our own blank cursor.
-				XUndefineCursor(dpy, win);
-				return;
-			}
-			//vars to make blank cursor
-			Pixmap blank;
-			XColor dummy;
-			char data[1] = {0};
-			Cursor cursor;
-			//make a blank cursor
-			blank = XCreateBitmapFromData (dpy, win, data, 1, 1);
-			if (blank == None)
-				std::cout << "error: out of memory." << std::endl;
-			cursor = XCreatePixmapCursor(dpy, blank, blank, &dummy, &dummy, 0, 0);
-			XFreePixmap(dpy, blank);
-			//this makes you the cursor. then set it using this function
-			XDefineCursor(dpy, win, cursor);
-			//after you do not need the cursor anymore use this function.
-			//it will undo the last change done by XDefineCursor
-			//(thus do only use ONCE XDefineCursor and then XUndefineCursor):
-		}
-} x11;
-
-static int colors = 255;
-int getColors(){
-        return colors;
-}
-
-static bool faded = 0;
-static bool fadeOutBegin = 0;
-extern void toggleMenu();
-bool menuFadedOut(){
-        return faded;
-}
-
-void menuSound(){
-        play_sound(gl.mgSound);
-        sleep(1);
-        play_sound(gl.sfSound);
-}
-
-void beginFade(){
-        fadeOutBegin=1;
-	std::thread td(menuSound);
-	td.detach();
-}
-
-void genBH(GLuint texture, int x, int y, int size){
-        int w = size;
-        //int h = 150;
-        glPushMatrix();
-        glTranslatef(x,y,0);
-        glBindTexture(GL_TEXTURE_2D, texture);
-        glEnable(GL_ALPHA_TEST);
-        glAlphaFunc(GL_GREATER, 0.0f);
-        glColor4ub(colors,colors,colors,255);
-        glBegin(GL_QUADS);
-        glTexCoord2f(0.0f, 1.0f); glVertex2i(-w,-w);
-        glTexCoord2f(0.0f, 0.0f); glVertex2i(-w, w);
-        glTexCoord2f(1.0f, 0.0f); glVertex2i( w, w);
-        glTexCoord2f(1.0f, 1.0f); glVertex2i( w,-w);
-        glEnd();
-        glPopMatrix();
-        glBindTexture(GL_TEXTURE_2D, 0);
-
-}
-
-bool fadeBegin(){
-        return fadeOutBegin;
-}
-
-void setColors(int x){
-        colors=x;
-}
-void genTitleScreen(GLuint texture,GLuint texture2, int x, int y){
-        int w = 800;
-        //int h = 150;
-        glPushMatrix();
-        glTranslatef(x,y+500,0);
-        glBindTexture(GL_TEXTURE_2D, texture);
-        glEnable(GL_ALPHA_TEST);
-        glAlphaFunc(GL_GREATER, 0.0f);
-        glColor4ub(colors,colors,colors,255);
-        glBegin(GL_QUADS);
-        glTexCoord2f(0.0f, 1.0f); glVertex2i(-w,-w);
-        glTexCoord2f(0.0f, 0.0f); glVertex2i(-w, w);
-        glTexCoord2f(1.0f, 0.0f); glVertex2i( w, w);
-        glTexCoord2f(1.0f, 1.0f); glVertex2i( w,-w);
-        glEnd();
-        glPopMatrix();
-        glBindTexture(GL_TEXTURE_2D, 0);
-
-
-        w = 350;
-        glPushMatrix();
-        glTranslatef(x,y,0);
-        glBindTexture(GL_TEXTURE_2D, texture2);
-        glEnable(GL_ALPHA_TEST);
-        glAlphaFunc(GL_GREATER, 0.0f);
-        glColor4ub(colors,colors,colors,255);
-        glBegin(GL_QUADS);
-        glTexCoord2f(0.0f, 1.0f); glVertex2i(-w,-w);
-        glTexCoord2f(0.0f, 0.0f); glVertex2i(-w, w);
-        glTexCoord2f(1.0f, 0.0f); glVertex2i( w, w);
-        glTexCoord2f(1.0f, 1.0f); glVertex2i( w,-w);
-        glEnd();
-        glPopMatrix();
-        glBindTexture(GL_TEXTURE_2D, 0);
-
-        if(fadeOutBegin==1&&colors>0){
-                colors-=1;
-		genBH(gl.bhTexture, 1200,1200,350);
-		genBH(gl.bhTexture, 2000,1600,350);
-		genBH(gl.bhTexture, 1800,1900,350);
-		genBH(gl.bhTexture, 1300,2000,350);
-		genBH(gl.bhTexture, 1500,2500,350);
-        }
-        if(colors==0){
-                faded =1;
-                toggleMenu();
-                colors=255;
-
-        }
-}
-
-
+}; //gl
